Replace MapVariant with MenuProps in multiple files
Replaced MapVariant with MenuProps in several files to streamline code and improve performance. The MapVariant enum was replaced with string props in the soil model and multiple uses of MapVariant were switched to MenuProps in the respective files. The refreshing metadata endpoint was also updated in the network/meta.ts file.